+++ src/Tests/Analyzers.Tests/RCS1014UseImplicitlyTypedArrayTests.cs
@@ -114,7 +114,7 @@ class B : A
         await VerifyDiagnosticAndFixAsync(@"
 class C
 {
-    string[][] _f = new [|string[]|][]
+    string[][] _f = new [|string|][][]
     {
         new[] { """" },
     };
@@ -180,5 +180,20 @@ class C
     }
 }
 ", options: Options.AddConfigOption(ConfigOptionKeys.ArrayCreationTypeStyle, ConfigOptionValues.ArrayCreationTypeStyle_ImplicitWhenTypeIsObvious));
+    }
+
+    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.UseExplicitlyOrImplicitlyTypedArray)]
+    public async Task TestNoDiagnostic_UseVarInsteadOfObjectCreation()
+    {
+        await VerifyNoDiagnosticAsync(@"
+class C
+{
+    void M()
+    {
+        var items = new string[] { string.Empty };
+    }
+}
+", options: Options.AddConfigOption(ConfigOptionKeys.ArrayCreationTypeStyle, ConfigOptionValues.ArrayCreationTypeStyle_Implicit)
+            .AddConfigOption(ConfigOptionKeys.UseVarInsteadOfImplicitObjectCreation, true));
     }
 }